Almost Original Source
The project was originally made in MonoGame 3.5, not 3.6. Also, I ran my code style tool through this project, so the code might not aesthetically look the same as the original. Also, content importing is slightly more automated than before. The game functions the same as the original, including the barely-working collision code.